refactored run command

+++ src/abstractions/ExpressConsensusNode.cs
@@ -5,13 +5,13 @@ namespace Neo.Express.Abstractions
     public class ExpressConsensusNode
     {
         [JsonProperty("tcp-port")]
-        public long TcpPort { get; set; }
+        public ushort TcpPort { get; set; }
 
         [JsonProperty("ws-port")]
-        public long WebSocketPort { get; set; }
+        public ushort WebSocketPort { get; set; }
 
         [JsonProperty("rpc-port")]
-        public long RpcPort { get; set; }
+        public ushort RpcPort { get; set; }
 
         [JsonProperty("wallet")]
         public ExpressWallet Wallet { get; set; }
